+++ srcs/handle_u.c
@@ -6,18 +6,18 @@
 /*   By: msakwins <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/01 15:46:28 by msakwins          #+#    #+#             */
-/*   Updated: 2017/05/31 20:11:35 by msakwins         ###   ########.fr       */
+/*   Updated: 2017/06/05 20:10:38 by msakwins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-size_t		handle_u(va_list argl, t_modif *modi)
+int			handle_u(va_list argl, t_modif *modi)
 {
 	uintmax_t			nb;
-	size_t				len;
-	size_t				nblen;
-	size_t				neg;
+	int					len;
+	int					nblen;
+	int					neg;
 
 	neg = 0;
 	len = 0;
